backend fixed and verification from supervisor added

+++ ChangeRequestApproval/src/main/java/com/esewa/restchangerequestapproval/requestChange/repo/RequestRepo.java
@@ -3,5 +3,6 @@ package com.esewa.restchangerequestapproval.requestChange.repo;
 import com.esewa.restchangerequestapproval.requestChange.entity.ChangeRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestRepo extends JpaRepository<ChangeRequest, Long> {
+public interface RequestRepo extends JpaRepository<ChangeRequest, Integer> {
+
 }
